refactored into multiple files (for each sensor)
