+++ src/main.cpp
@@ -1,171 +1,15 @@
 #include <Arduino.h>
-#include "MQ135.h"
-
-// MQ135 gas sensor
-//
-// Datasheet can be found here: https://www.olimex.com/Products/Components/Sensors/SNS-MQ135/resources/SNS-MQ135.pdf
-//
-// Application
-// They are used in air quality control equipments for buildings/offices, are suitable for detecting of NH3, NOx, alcohol, Benzene, smoke, CO2, etc
-//
-// Original creator of this library: https://github.com/GeorgK/MQ135
-
-#define PIN_MQ135 A0
-MQ135 mq135_sensor = MQ135(PIN_MQ135);
-
-float DEFAULT_TEMPERATURE = 21.0;
-float DEFAULT_HUMIDITY = 25.0;
-
-void MQ135_loop(float temperature, float humidity);
-
-// DHT Temperature & Humidity Sensor
-
-// REQUIRES the following Arduino libraries:
-// - DHT Sensor Library: https://github.com/adafruit/DHT-sensor-library
-// - Adafruit Unified Sensor Lib: https://github.com/adafruit/Adafruit_Sensor
-
-#include <Adafruit_Sensor.h>
-#include <DHT.h>
-#include <DHT_U.h>
-
-#define DHTPIN D2 // Digital pin connected to the DHT sensor
-// Feather HUZZAH ESP8266 note: use pins 3, 4, 5, 12, 13 or 14 --
-// Pin 15 can work but DHT must be disconnected during program upload.
-
-// Uncomment the type of sensor in use:
-// #define DHTTYPE    DHT11     // DHT 11
-#define DHTTYPE DHT22 // DHT 22 (AM2302)
-// #define DHTTYPE    DHT21     // DHT 21 (AM2301)
-
-// See guide for details on sensor wiring and usage:
-//   https://learn.adafruit.com/dht/overview
-
-DHT_Unified dht(DHTPIN, DHTTYPE);
-
-uint32_t delayMS;
-
-struct result
-{
-    float temperature;
-    float humidity;
-};
-
-void DHT_setup();
-result DHT_loop();
+#include "mq.h"
+#include "dht22.h"
 
 void setup()
 {
     Serial.begin(9600);
-    delay(1000); // Let serial console settle so we can see startup messages
     DHT_setup();
 }
 
 void loop()
 {
-    auto result = DHT_loop();
+    DHT_Result result = DHT_loop();
     MQ135_loop(result.temperature, result.humidity);
-}
-
-void MQ135_loop(float temperature, float humidity)
-{
-    float rzero = mq135_sensor.getRZero();
-    float correctedRZero = mq135_sensor.getCorrectedRZero(temperature, humidity);
-    float resistance = mq135_sensor.getResistance();
-    float ppm = mq135_sensor.getPPM();
-    float correctedPPM = mq135_sensor.getCorrectedPPM(temperature, humidity);
-
-    Serial.print("MQ135 RZero: ");
-    Serial.print(rzero);
-    Serial.print("\t Corrected RZero: ");
-    Serial.print(correctedRZero);
-    Serial.print("\t Resistance: ");
-    Serial.print(resistance);
-    Serial.print("\t PPM: ");
-    Serial.print(ppm);
-    Serial.print("\t Corrected PPM: ");
-    Serial.print(correctedPPM);
-    Serial.println("ppm");
-
-    delay(300);
-}
-
-void DHT_setup()
-{
-    // Initialize device.
-    dht.begin();
-    Serial.println(F("DHTxx Sensor Setup"));
-    // Print temperature sensor details.
-    sensor_t sensor;
-    dht.temperature().getSensor(&sensor);
-    Serial.println(F("------------------------------------"));
-    Serial.println(F("Temperature Sensor"));
-    Serial.print(F("Sensor Type: "));
-    Serial.println(sensor.name);
-    Serial.print(F("Driver Ver:  "));
-    Serial.println(sensor.version);
-    Serial.print(F("Unique ID:   "));
-    Serial.println(sensor.sensor_id);
-    Serial.print(F("Max Value:   "));
-    Serial.print(sensor.max_value);
-    Serial.println(F("°C"));
-    Serial.print(F("Min Value:   "));
-    Serial.print(sensor.min_value);
-    Serial.println(F("°C"));
-    Serial.print(F("Resolution:  "));
-    Serial.print(sensor.resolution);
-    Serial.println(F("°C"));
-    Serial.println(F("------------------------------------"));
-    // Print humidity sensor details.
-    dht.humidity().getSensor(&sensor);
-    Serial.println(F("Humidity Sensor"));
-    Serial.print(F("Sensor Type: "));
-    Serial.println(sensor.name);
-    Serial.print(F("Driver Ver:  "));
-    Serial.println(sensor.version);
-    Serial.print(F("Unique ID:   "));
-    Serial.println(sensor.sensor_id);
-    Serial.print(F("Max Value:   "));
-    Serial.print(sensor.max_value);
-    Serial.println(F("%"));
-    Serial.print(F("Min Value:   "));
-    Serial.print(sensor.min_value);
-    Serial.println(F("%"));
-    Serial.print(F("Resolution:  "));
-    Serial.print(sensor.resolution);
-    Serial.println(F("%"));
-    Serial.println(F("------------------------------------"));
-    // Set delay between sensor readings based on sensor details.
-    delayMS = sensor.min_delay / 1000;
-}
-
-result DHT_loop()
-{
-    // Delay between measurements.
-    delay(delayMS);
-    // Get temperature event and print its value.
-    sensors_event_t event;
-    dht.temperature().getEvent(&event);
-    if (isnan(event.temperature))
-    {
-        Serial.println(F("Error reading temperature!"));
-    }
-    else
-    {
-        Serial.print(F("Temperature: "));
-        Serial.print(event.temperature);
-        Serial.println(F("°C"));
-    }
-    // Get humidity event and print its value.
-    dht.humidity().getEvent(&event);
-    if (isnan(event.relative_humidity))
-    {
-        Serial.println(F("Error reading humidity!"));
-    }
-    else
-    {
-        Serial.print(F("Humidity: "));
-        Serial.print(event.relative_humidity);
-        Serial.println(F("%"));
-    }
-    return result{event.temperature, event.relative_humidity};
 }
