+++ src/mq.cpp
@@ -1,0 +1,28 @@
+#include "mq.h"
+
+MQ135 mq135_sensor = MQ135(PIN_MQ135);
+
+MQ135_Result MQ135_loop(float temperature, float humidity)
+{
+    float rzero = mq135_sensor.getRZero();
+    float correctedRZero = mq135_sensor.getCorrectedRZero(temperature, humidity);
+    float resistance = mq135_sensor.getResistance();
+    float ppm = mq135_sensor.getPPM();
+    float correctedPPM = mq135_sensor.getCorrectedPPM(temperature, humidity);
+
+    Serial.print("MQ135 RZero: ");
+    Serial.print(rzero);
+    Serial.print("\t Corrected RZero: ");
+    Serial.print(correctedRZero);
+    Serial.print("\t Resistance: ");
+    Serial.print(resistance);
+    Serial.print("\t PPM: ");
+    Serial.print(ppm);
+    Serial.print("\t Corrected PPM: ");
+    Serial.print(correctedPPM);
+    Serial.println("ppm");
+
+    delay(300);
+
+    return MQ135_Result{rzero, correctedRZero, resistance, ppm, correctedPPM};
+}
